+++ packages/cli/src/utils/index.ts
@@ -1,7 +1,7 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './appBridgeVersion';
-export * from './compiler/compiler';
+export * from './compiler/compilerOptions';
 export * from './configuration';
 export * from './date';
 export * from './file';
